# Benchmark some wallet code, fix a serious bug, and speed things up