--- a/packages/server-wallet/e2e-test/payer/client.ts
+++ b/packages/server-wallet/e2e-test/payer/client.ts
@@ -83,16 +83,18 @@ export default class PayerClient {
   }
 
   public async makePayment(channelId: string): Promise<void> {
-    const channel = await this.getChannel(channelId);
+    const channel = await time(`get channel ${channelId}`, async () => this.getChannel(channelId));
 
     // Assuming MessageQueued inside the outbox
     const {
       outbox: [{params}],
-    } = await this.wallet.updateChannel(channel);
+    } = await time(`update ${channelId}`, async () => this.wallet.updateChannel(channel));
 
-    const reply = await this.messageReceiverAndExpectReply((params as WireMessage).data as Message);
+    const reply = await time(`send message ${channelId}`, async () =>
+      this.messageReceiverAndExpectReply((params as WireMessage).data as Message)
+    );
 
-    await this.wallet.pushMessage(reply);
+    await time(`push message ${channelId}`, async () => this.wallet.pushMessage(reply));
   }
 
   public emptyMessage(): Promise<Message> {
@@ -105,5 +107,18 @@ export default class PayerClient {
   private async messageReceiverAndExpectReply(message: Message): Promise<Message> {
     const {data: reply} = await axios.post(this.receiverHttpServerURL + '/inbox', {message});
     return reply;
+  }
+}
+
+// eslint-disable-next-line no-process-env
+const TIME = !!process.env.TIMING_METRICS;
+async function time<T>(label: string, cb: () => Promise<T>): Promise<T> {
+  if (TIME) {
+    console.time(label);
+    const result = await cb();
+    console.timeEnd(label);
+    return result;
+  } else {
+    return await cb();
   }
 }