--- a/packages/server-wallet/e2e-test/receiver/controller.ts
+++ b/packages/server-wallet/e2e-test/receiver/controller.ts
@@ -23,7 +23,7 @@ export default class ReceiverController {
 
     const {
       channelResults: [channelResult],
-    } = await this.wallet.pushMessage(message);
+    } = await time('controller: push message', async () => this.wallet.pushMessage(message));
 
     if (!signedStates || signedStates?.length === 0) {
       return {
@@ -33,11 +33,26 @@ export default class ReceiverController {
     } else {
       const {
         outbox: [messageToSendToPayer],
-      } = await (channelResult.status === 'proposed'
-        ? this.wallet.joinChannel
-        : this.wallet.updateChannel)(channelResult);
+      } = await time('controller: react', async () =>
+        (channelResult.status === 'proposed' ? this.wallet.joinChannel : this.wallet.updateChannel)(
+          channelResult
+        )
+      );
 
       return (messageToSendToPayer.params as WireMessage).data as Message;
     }
+  }
+}
+
+// eslint-disable-next-line no-process-env
+const TIME = !!process.env.TIMING_METRICS;
+async function time<T>(label: string, cb: () => Promise<T>): Promise<T> {
+  if (TIME) {
+    console.time(label);
+    const result = await cb();
+    console.timeEnd(label);
+    return result;
+  } else {
+    return await cb();
   }
 }