--- a/packages/server-wallet/e2e-test/receiver/controller.ts
+++ b/packages/server-wallet/e2e-test/receiver/controller.ts
@@ -2,6 +2,7 @@ import {Message as WireMessage} from '@statechannels/wire-format';
 import {Message, makeDestination} from '@statechannels/wallet-core';
 import {Participant} from '@statechannels/client-api-schema';
 
+import walletConfig from '../../src/config';
 import {bob} from '../../src/wallet/__test__/fixtures/signing-wallets';
 import {Wallet} from '../../src/wallet';
 
@@ -45,9 +46,8 @@ export default class ReceiverController {
 }
 
 // eslint-disable-next-line no-process-env
-const TIME = !!process.env.TIMING_METRICS;
 async function time<T>(label: string, cb: () => Promise<T>): Promise<T> {
-  if (TIME) {
+  if (walletConfig.timingMetrics) {
     console.time(label);
     const result = await cb();
     console.timeEnd(label);