--- a/packages/server-wallet/src/metrics.ts
+++ b/packages/server-wallet/src/metrics.ts
@@ -1,12 +1,12 @@
+import walletConfig from './config';
+
 export const timerFactory = (prefix: string) => async <T>(
   label: string,
   cb: () => Promise<T>
 ): Promise<T> => time(`${prefix}: ${label}`, cb);
 
-// eslint-disable-next-line no-process-env
-const TIME = !!process.env.TIMING_METRICS;
 async function time<T>(label: string, cb: () => Promise<T>): Promise<T> {
-  if (TIME) {
+  if (walletConfig.timingMetrics) {
     console.time(label);
     const result = await cb();
     console.timeEnd(label);