This adds a naive `timerFactory` which lets us [time](https://github.com/statechannels/statechannels/compare/ags/wallet-benchmarks?expand=1#diff-da530680787955aca14faaf1899bcce0R87) some async, and slow sync function calls.

By [inspecting the output](https://github.com/statechannels/statechannels/commit/a65e6001b59b373cad67982938ead7be0c84f057) and auditing the code, I removed a number of database queries.

I also used [`fastSignState`](https://github.com/statechannels/statechannels/commit/d1bf1f3e75ef1e8d9a8bb88c000196c55d6fba4d) implemented by @kerzhner , which was a huge time saver in the stress test:
- [before d1bf1f3e75ef1e8d9a8bb88c000196c55d6fba4d](https://app.circleci.com/pipelines/github/statechannels/statechannels/8670/workflows/995f686f-d3f7-4c08-a3db-8166ef22270a/jobs/40855)
- [after d1bf1f3e75ef1e8d9a8bb88c000196c55d6fba4d](https://app.circleci.com/pipelines/github/statechannels/statechannels/8671/workflows/037cf4d9-0d52-4196-9362-1f5008fb2288/jobs/40862)

Finally, I identified a [serious bug](https://github.com/statechannels/statechannels/commit/ce077a490da229364da286fdfce8f22e87fd416b) which would create a deadlock when too many update channel calls are made simultaenously.

# Results
## On master (d6929e0)
```
monorepo/packages/server-wallet on  master [$!?] is 📦 v0.3.0 via ⬢ v12.16.3 on ☁️  us-west-1 took 2s 
❯ ts-node benchmarks/update-channel.benchmark.ts
[@statechannels/devtools] NODE_ENV is undefined — setting to "development"
serial x 100: 2597.107ms
concurrent x 100: 0.007ms
```

Stress test on master (with 100 channels)
```
┌─────────────────────────────────────┬────────────────────┬───────────────────┬───────────────────┐
│ Action                              │ Min (MS)           │ Max (MS)          │ Avg (MS)          │
├─────────────────────────────────────┼────────────────────┼───────────────────┼───────────────────┤
│ Individual makePayment call         │ 1561.1062720008194 │ 6970.285530000925 │ 3949.791616135605 │
├─────────────────────────────────────┼────────────────────┼───────────────────┼───────────────────┤
│ 25 consecutive calls of makePayment │ 97704.9870209992   │ 99348.86164898798 │ 98744.79040339016 │
└─────────────────────────────────────┴────────────────────┴───────────────────┴───────────────────┘
```


## This branch (d1bf1f3)
```
monorepo/packages/server-wallet on  ags/wallet-benchmarks [$?] is 📦 v0.3.0 via ⬢ v12.16.3 on ☁️  us-west-1 
❯ ts-node benchmarks/update-channel.benchmark.ts
[@statechannels/devtools] NODE_ENV is undefined — setting to "development"
serial x 100: 1449.653ms
concurrent x 100: 1392.089ms
```

[Stress test](https://app.circleci.com/pipelines/github/statechannels/statechannels/8671/workflows/037cf4d9-0d52-4196-9362-1f5008fb2288/jobs/40862/parallel-runs/0/steps/0-104) (also 100 channels)

```
┌─────────────────────────────────────┬────────────────────┬───────────────────┬───────────────────┐
│ Action                              │ Min (MS)           │ Max (MS)          │ Avg (MS)          │
├─────────────────────────────────────┼────────────────────┼───────────────────┼───────────────────┤
│ Individual makePayment call         │ 1481.4771390110254 │ 4240.864399001002 │ 2173.318449879604 │
├─────────────────────────────────────┼────────────────────┼───────────────────┼───────────────────┤
│ 25 consecutive calls of makePayment │ 53470.81020399928  │ 55041.33449099958 │ 54332.96124699011 │
└─────────────────────────────────────┴────────────────────┴───────────────────┴───────────────────┘
```

# Remaining improvements
- [ ] Repeat for `pushMessage`
- [x] Getting the address from an ethers wallet is slow
- [ ] verifying signatures with ethers is slow
- [ ] hashing states with ethers is slow